Move json_schema into seperate files
