+++ ert_storage/json_schema/__init__.py
@@ -1,0 +1,3 @@
+from .ensemble import EnsembleIn, EnsembleOut
+from .record import RecordOut
+from .experiment import ExperimentIn, ExperimentOut
