+++ ert_storage/json_schema/ensemble.py
@@ -1,0 +1,17 @@
+from typing import List
+from pydantic import BaseModel
+
+
+class _Ensemble(BaseModel):
+    pass
+
+
+class EnsembleIn(_Ensemble):
+    parameters: List[str]
+
+
+class EnsembleOut(_Ensemble):
+    id: int
+
+    class Config:
+        orm_mode = True
